[#TAPO-11] Implemented GameObject-Component system
This commit introduces a big change on the technical side of the game - instead of creating a new class for each type of tile we now have components which can implement custom functionality and allow us to reuse previously implemented features without copy-pasting anything.

This commit only integrates the system with the engine, blocks have not been rewritten to components, yet.